training/requirements.txt
anthropic>=0.5.0
openai>=1.0.0
python-dotenv>=1.0.0
pymongo>=4.5.0
bson>=0.5.10
tiktoken>=0.5.1
numpy>=1.24.0